restructured rdf_page_data
