+++ lib/jekyll-rdf.rb
@@ -45,6 +45,7 @@ require 'jekyll/filters/rdf_property'
 require 'jekyll/filters/rdf_collection'
 require 'jekyll/filters/rdf_container'
 require 'jekyll/rdf_main_generator'
+require 'jekyll/rdf_page_helper'
 require 'jekyll/rdf_page_data'
 require 'jekyll/rdf_class_extraction'
 require 'jekyll/rdf_template_mapper'
